+++ crates/tycho-test/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod execution;
 pub mod rpc_tools;
+pub mod token_prices;
 pub mod validation;
 pub use rpc_tools::RPCTools;